requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
chatarena @ git+https://github.com/chatarena/chatarena@2aeb4fc7bc59db38076a7f4b317bc3569856ecf2
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
jiter==0.7.0
markdown-it-py==2.2.0
mdurl==0.1.2
numpy==2.1.3
openai==1.54.0
packaging==24.1
pillow==11.0.0
prompt-toolkit==3.0.38
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.3.3
safetensors==0.4.5
sniffio==1.3.1
tenacity==8.2.2
tokenizers==0.19.1
tqdm==4.66.6
transformers==4.41.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
